Diffusion_model_and_other_thesis_files/diffusion_model_requirements.txt
torch==2.3.1+cu121
torchvision==0.18.1+cu121
Python==3.9.21
huggingface_hub==0.16.4
numpy==1.24.2
safetensors==0.5.3
scipy==1.13.1
timm==0.9.12
scikit-image==0.18.3
torcheval==0.0.7
tokenizer==0.13.3
torchmetrics==1.5.2
transformers==4.33.3
xformers==0.0.27
accelerate==0.31.0
diffusers==0.23.0
einops==0.8.1
fairscale==0.4.13
loralib==0.1.2
pandas==2.0.1
pydantic==1.10.6
pytorch-lightning==2.3.3
opencv==4.10.0.84
